refactor: ui 유틸리티 패키지 분리 및 번들 키 상수화
