+++ app/src/main/java/woowacourse/payments/ui/CardUiModel.kt
@@ -3,6 +3,8 @@ package woowacourse.payments.ui
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 import woowacourse.payments.domain.Card
+import woowacourse.payments.ui.util.formatCardNumber
+import woowacourse.payments.ui.util.formatExpired
 
 @Parcelize
 data class CardUiModel(
