+++ app/src/main/java/woowacourse/payments/ui/util/BundleKeys.kt
@@ -1,0 +1,9 @@
+package woowacourse.payments.ui.util
+
+object BundleKeys {
+    const val CARD_NUMBER_KEY = "cardNumber"
+    const val EXPIRED_KEY = "expired"
+    const val CARD_OWNER_KEY = "cardOwner"
+    const val PASSWORD_KEY = "password"
+    const val VALIDATION_ERROR_KEY = "showValidationError"
+}
